Reestructuracion

+++ hartz_backend/src/main/java/org/hartz/hartz_backend/user/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package org.hartz.hartz_backend.user.repository;
 
-import org.hartz.hartz_backend.user.User;
+import org.hartz.hartz_backend.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByUsername(String username);
     boolean existsByEmail(String email);
